% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNetworkBaseUrl.R
\name{getNetworkBaseUrl}
\alias{getNetworkBaseUrl}
\title{getNetworkBaseUrl}
\usage{
getNetworkBaseUrl(network = c("mainnet", "previewnet", "testnet")[3])
}
\arguments{
\item{network}{Character. The name of the Hedera network for which to return 
the mirror node base URL. Must be one of "mainnet", "previewnet" or "testnet".}
}
\description{
Returns the mirror node base URL for a given Hedera network.
}
