% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTokenBalances.R
\name{getTokenBalances}
\alias{getTokenBalances}
\title{getTokenBalances}
\usage{
getTokenBalances(
  tokenId,
  accountId = NULL,
  balanceFilter = "gt:0",
  atTimeStamp = NULL,
  limit = 100,
  order = c("asc", "desc")[1],
  network = c("mainnet", "previewnet", "testnet")[3],
  asDf = FALSE
)
}
\arguments{
\item{tokenId}{Character. The Hedera ID for the token, e.g., 0.0.24315.}

\item{accountId}{Character. Optional. The account for which the balance
should be retrieved. If not provided, the balances for all associated accounts 
will be retrieved.}

\item{balanceFilter}{Character. Optional. The filter to apply to accounts, e.g.,
'gt:1000' will only return the balances for accounts with more than 1000
tokenId tokens. Defaults to 'gt:0'.}

\item{atTimeStamp}{Numeric. Optional. The timestamp, in seconds.nanoseconds,
for which the balances should be retrieved.}

\item{limit}{Integer. Optional. The maximum number of results to return. 
Defaults to 100.}

\item{order}{Character. Optional. 'asc' will sort the accounts in increasing
order of balance; 'desc' will sort the accounts in decreasing order of 
balance. Defaults to 'asc'.}

\item{network}{Character. Optional. The network to connect to. Defaults to
'testnet'.}

\item{asDf}{Logical. Optional. If TRUE, the function will return the results
as a data frame. Defaults to FALSE, which will return the results as a list.}
}
\value{
A list of account balances for the given token.
}
\description{
Gets the balances across all associated accounts for a given token.
   This represents the token supply distribution across the network.
}
