% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMessage.R
\name{getMessage}
\alias{getMessage}
\title{getMessage}
\usage{
getMessage(
  messageId,
  network = c("mainnet", "previewnet", "testnet")[3],
  decode = FALSE
)
}
\arguments{
\item{messageId}{The ID of the message to retrieve, e.g., "1709313529.822677003".}

\item{network}{Character. Which Hedera network to use. Must be one of 
"mainnet", "previewnet" or "testnet".}

\item{decode}{Logical. If TRUE, will decode the base64-encoded memo field of
the message; if FALSE, will leave the memo field as is. Default is FALSE.}
}
\description{
Retrieves a Hedera message from the network.
}
