% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNfts.R
\name{getNfts}
\alias{getNfts}
\title{getNfts}
\usage{
getNfts(
  tokenId,
  accountId = NULL,
  serialNo = NULL,
  limit = 100,
  order = c("asc", "desc")[1],
  network = c("mainnet", "previewnet", "testnet")[3],
  asDf = FALSE
)
}
\arguments{
\item{limit}{Integer. Optional. The maximum number of results to return. 
Defaults to 100.}

\item{order}{Character. Optional. 'asc' will sort the results in ascending
order of serial number; 'desc' will sort the results in descending order of
serial number. Defaults to 'asc'.}

\item{network}{Character. Optional. The network to connect to. Defaults to
'testnet'.}

\item{asDf}{Logical. Optional. If TRUE, the function will return the results
as a data frame. Defaults to FALSE, which will return the results as a list.}
}
\description{
Returns metadata about the instances of a given NFT, identified by token ID.
   Can optionally filter by Hedera account ID and/or NFT serial number.
}
